requirements_depth.txt
matplotlib==3.5.1
numpy==1.22.2
pandas==2.1.4
onnxruntime==1.12.0
onnxruntime-gpu==1.12.0
opencv-python==4.5.4.60
pyinstaller==5.6.2
scikit-learn==1.1.1
tqdm==4.64.0
pillow
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
transformers==4.34.0
hydra-core==1.3.2
jupyter
prettytable
wildlifeml @ git+https://github.com/slds-lmu/wildlife-ml.git@e46a854e5bcdac88538667c315c00c16601d29e8